restructure server folder
