+++ app.js
@@ -2,4 +2,4 @@
 // Initialize server
 //////////////////////////////////////////////////////////////////////////////////////////////
 
-require('./src/server/app.js')
+require('./src/server/src/app.js')
