+++ src/server/src/app.js
@@ -25,7 +25,7 @@ const MongoStore = require('connect-mongo');
 // Load configuration
 // ////////////////////////////////////////////////////////////////////////////////////////////////
 
-global.config = require('./../config')
+global.config = require('./config')
 
 // ////////////////////////////////////////////////////////////////////////////////////////////////
 // Express application
