+++ src/server/src/middleware/authenticated.js
