+++ src/server/src/middleware/cleanup.js
