+++ src/tests/server/documents/utils.js
@@ -1,4 +1,4 @@
-const utils = require('../../../server/documents/utils');
+const utils = require('../../../server/src/documents/utils');
 const assert = require('assert');
 
 describe('documents:utils', () => {
