+++ src/tests/server/services/org.js
@@ -5,10 +5,10 @@ const assert = require('assert')
 const sinon = require('sinon')
 
 //model
-const Org = require('../../../server/documents/org').Org
+const Org = require('../../../server/src/documents/org').Org
 
 // service under test
-const org = require('../../../server/services/org')
+const org = require('../../../server/src/services/org')
 
 // test data
 const testData = require('../testData').data
