+++ src/tests/server/webhooks/pull_request.js
@@ -4,19 +4,19 @@ const assert = require('assert')
 const sinon = require('sinon')
 
 // services
-const pullRequest = require('../../../server/services/pullRequest')
-const status = require('../../../server/services/status')
-const cla = require('../../../server/services/cla')
-const repoService = require('../../../server/services/repo')
-const orgService = require('../../../server/services/org')
-const logger = require('../../../server/services/logger')
+const pullRequest = require('../../../server/src/services/pullRequest')
+const status = require('../../../server/src/services/status')
+const cla = require('../../../server/src/services/cla')
+const repoService = require('../../../server/src/services/repo')
+const orgService = require('../../../server/src/services/org')
+const logger = require('../../../server/src/services/logger')
 
-const config = require('../../../config')
+const config = require('../../../server/src/config')
 
-const User = require('../../../server/documents/user').User
+const User = require('../../../server/src/documents/user').User
 
 // webhook under test
-const webhook = require('../../../server/webhooks/pull_request')
+const webhook = require('../../../server/src/webhooks/pull_request')
 
 function pull_request(req, res) {
     if (webhook.accepts(req)) {
